% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{choose_font_colour}
\alias{choose_font_colour}
\title{Choose font colour}
\usage{
choose_font_colour(bg, light = "#ffffff", dark = "#000000", threshold = 170)
}
\arguments{
\item{bg}{Background}

\item{light}{The light text colour}

\item{dark}{The dark text colour}

\item{threshold}{The threshold for switching}
}
\value{
hex code
}
\description{
Choose font colour
}
