% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyedropper.R
\name{extract_pal}
\alias{extract_pal}
\title{Extracts palette from image}
\usage{
extract_pal(n, img_path = NULL)
}
\arguments{
\item{img_path}{}
}
\value{
Returns a character vector of hex codes
}
\description{
The image is read in using \code{magick}, converted to RGB and clustered using kmeans. The user
must specify the number of clusters. The cluster centroids become the palette values.
}
\examples{
\dontrun{
path <- "https://colorpalettes.net/wp-content/uploads/2015/05/cvetovaya-palitra-1781.png"
extract_pal(path)
}
}
