% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-palette.R
\name{generate_palette}
\alias{generate_palette}
\title{Generate palette}
\usage{
generate_palette(n, prompt, name = NULL)
}
\arguments{
\item{n}{Number of colours in the palette}

\item{prompt}{Prompt}

\item{name}{Name for the palette. If left NULL it will}
}
\description{
Generate palette
}
\examples{
# x <- generate_palette(4, 'colourful balls in a chldrens ball pit', name = 'Ball pit')
}
