% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyedropper.R
\name{make_output}
\alias{make_output}
\title{Makes eyedroppers output}
\usage{
make_output(.pal, .img_path, .label)
}
\arguments{
\item{.pal}{Palette}

\item{.img_path}{Image path}

\item{.label}{Label}
}
\value{
ggplot object
}
\description{
Plots the palette and places the image and label over the top.
}
