% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{min_max}
\alias{min_max}
\title{Min-max scale function}
\usage{
min_max(x, a, b, a0 = NULL, b0 = NULL)
}
\arguments{
\item{x}{Vector of values}

\item{a}{Min}

\item{b}{Max}

\item{a0}{Min bound}

\item{b0}{Max bound}
}
\value{
numeric vector
}
\description{
Min-max scale function
}
