% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{modify_saturation}
\alias{modify_saturation}
\title{Modify the saturation of a colour palette}
\usage{
modify_saturation(cols, sat = 1.2)
}
\arguments{
\item{cols}{Vector of colours}

\item{sat}{Factor to adjust the saturation}
}
\description{
Modify the saturation of a colour palette
}
