% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pencil-case.R
\name{palette}
\alias{palette}
\title{Easily get a palette from the pencil case}
\usage{
palette(id)
}
\arguments{
\item{id}{Number}
}
\description{
Easily get a palette from the pencil case
}
\examples{
palette(1)
}
