% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{paste_pal_code}
\alias{paste_pal_code}
\title{Copy + Pasteable palette vector}
\usage{
paste_pal_code(pal, label = NULL)
}
\arguments{
\item{pal}{Palette vector}

\item{label}{Label}
}
\value{
a message
}
\description{
Prints a message to console so you can easily copy and paste the palette
}
