% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pastey}
\alias{pastey}
\title{Copy + Pastable palette vector}
\usage{
pastey(.pal, .label = NULL)
}
\arguments{
\item{.pal}{Palette vector}

\item{.label}{Label}
}
\value{
a message
}
\description{
Prints a message to console so you can easily copy and paste the palette
}
