% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-palette.R
\name{set_openai_key}
\alias{set_openai_key}
\title{Set OpenAI API key}
\usage{
set_openai_key(key)
}
\arguments{
\item{key}{Key from OpenAI account}
}
\value{
Nothing
}
\description{
Set OpenAI API key
}
