% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyedropper.R
\name{show_pal}
\alias{show_pal}
\title{Show palette}
\usage{
show_pal(pal)
}
\arguments{
\item{pal}{Palette. Vector of hex codes}
}
\value{
ggplot
}
\description{
Plots the supplied palette for quick inspection
}
\examples{
pal <- c('#57364e', '#566f1b', '#97a258', '#cac58b', '#dbedd5')
show_pal(pal)
}
