% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pencil-case.R
\name{show_pencil_case}
\alias{show_pencil_case}
\title{Shows all palettes in the pencil case}
\usage{
show_pencil_case(offset = 50)
}
\arguments{
\item{offset}{Controls the overlapping circles}
}
\value{
graphic of colour palettes
}
\description{
Shows all palettes in the pencil case
}
\examples{
show_pencil_case()
}
