% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_pal}
\alias{sort_pal}
\title{Manually sort a palette}
\usage{
sort_pal(pal, n = NULL, label = NULL, print_output = TRUE)
}
\arguments{
\item{pal}{Palette. Character vector of hex codes}

\item{n}{Number of colours to choose}

\item{label}{Label for the palette.}

\item{print_output}{Print output to console to easily copy and paste into your script.}
}
\value{
Character vector
}
\description{
The palette is displayed in the plotting window where you can click
the colours in the order you want to sort them. The sorted palette
will be returned. This saves you copy/pasting hex codes in your script.
}
\examples{
\dontrun{
pal <- sample(c('#57364e', '#566f1b', '#97a258', '#cac58b', '#dbedd5'))
sort_pal(pal)
}
}
