% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swatch}
\alias{swatch}
\title{Swatch of colour palette}
\usage{
swatch(
  pal,
  img = NULL,
  label = NULL,
  family = "Poppins",
  padding = 0,
  radius = 20,
  ncols = NULL,
  img_shadow = TRUE,
  shape = "original"
)
}
\arguments{
\item{pal}{Vector of colours}

\item{img}{Image address. Either local file or URL}

\item{label}{Label for the palette.}

\item{family}{Font family}

\item{padding}{To add whitespace to the top of image}

\item{radius}{The radius of the feature image. Choose 50 for a circle, less then 50 for rounded square/rectangle.}

\item{ncols}{Number of cols in a row}

\item{img_shadow}{Logical. Apply a shadow to the feature image output.}

\item{shape}{Either 'original' for the original image dimensions or "square"
to crop to square with rounded corners. A high enough radius will turn the square to a circle.}
}
\value{
html doc
}
\description{
Swatch of colour palette
}
\examples{
url <- "https://github.com/doehm/eyedroppeR/raw/main/dev/images/sunset-south-coast.jpg"
x <- extract_pal(4, url)
swatch(x$pal, url)
}
