% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyedropper.R
\name{extract_pal}
\alias{extract_pal}
\title{Extracts palette from an image}
\usage{
extract_pal(n, img_path, label = NULL, sort = "auto")
}
\arguments{
\item{n}{Number of colours to extract}

\item{img_path}{Path to image. If `NULL` the function will read from the clipboard}

\item{label}{Label for the palette.}

\item{sort}{Sort method. Either 'manual' or 'auto'}
}
\value{
Returns a character vector of hex codes
}
\description{
The image is read in using \code{magick}, converted to RGB and clustered using kmeans. The user
must specify the number of clusters. The cluster centroids become the palette values.
The function will ask the user to sort the palette values after clustering
}
\examples{
\dontrun{
path <- file.path(system.file(package = "eyedroppeR"), "images", "sunset.png")
extract_pal(8, path)
}
}
