% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyedropper.R
\name{eyedropper}
\alias{eyedropper}
\title{Eyedropper}
\usage{
eyedropper(n, img_path = NULL)
}
\arguments{
\item{n}{Number of colours to extract from the image}

\item{img_path}{Path to image. Can be local or from a URL. If left \code{NULL},
\code{eyedropper} will read the image address directly from the clipboard.}
}
\value{
A character vector of hex codes
}
\description{
Plots an image and allows the user to click on the image to
return the hex of the pixel. Can select multiple colours at
once by setting \code{n}.
}
\details{
Use \code{eyedropper} with the following steps:
\enumerate{
  \item{Find the image you want to pick colours from online.}
  \item{Right-click and 'copy image address'.}
  \item{Choose how many colours to pick e.g. \code{n = 5}.}
  \item{Run \code{pal <- eyedropper(n = 5)}. The function will read the copied address from the clipboard.}
  \item{Click 5 areas of the image. The image will be stretched to the borders of the window, but that's OK.}
  \item{Done! Copy the returned string and add it to you script and start using \code{pal}}
}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# image from https://colorpalettes.net/color-palette-1781/
path <- "https://colorpalettes.net/wp-content/uploads/2015/05/cvetovaya-palitra-1781.png"

# Run eyedropper and click on 5 colours
x <- eyedropper(n = 5, path)

x
\dontshow{\}) # examplesIf}
}
