% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyedropper.R
\name{eyedropper}
\alias{eyedropper}
\title{Eyedropper}
\usage{
eyedropper(
  n,
  img_path = NULL,
  inc_palette = TRUE,
  n_swatches = 24,
  print_output = TRUE,
  calibrate = FALSE,
  swatch_radius = 50
)
}
\arguments{
\item{n}{Number of colours to extract from the image}

\item{img_path}{Path to image. Can be local or from a URL. If left \code{NULL},
\code{eyedropper} will read the image address directly from the clipboard.}

\item{inc_palette}{Logical. If \code{TRUE} it will automatically extract a palette
first and then you can select the desired colours.}

\item{n_swatches}{Number of swatches to extract from the image prior to selecting colours.s}

\item{print_output}{Print output to console to easily copy and paste into your script.}

\item{calibrate}{Set to `TRUE` to calibrate the plot coordinates. Given the monitor
resolution, scaling, etc it can throw off the pixel selection. Runs but default the first
time the function is used.}

\item{swatch_radius}{Radius of the image for the swatch. Default 50 to make it a circle. Use 5 for rounded edges.}
}
\value{
A character vector of hex codes
}
\description{
Plots an image and allows the user to click on the image to
return the hex of the pixel. Can select multiple colours at
once by setting \code{n}.
}
\details{
Use \code{eyedropper} with the following steps:
\enumerate{
  \item{Find the image you want to pick colours from online.}
  \item{Right-click and 'copy image address'.}
  \item{Choose how many colours to pick e.g. \code{n = 5}.}
  \item{Run \code{pal <- eyedropper(n = 5, img_path = 'paste-image-path-here')}.}
  \item{Click 5 areas of the image. The image will be stretched to the borders of the window, but that's OK.}
  \item{Done! Copy the returned string / message and add it to you script and start using \code{pal}}
}

The first time the function is run it will initiate a calibration set. This is so the `y` coordinates are
scaled properly as this can depend on the monitors resolution, scaling, etc. In only takes a couple of seconds
and you only have to do it once.

Make sure you click as near as practicable to the top and bottom of the border of the windew within the dot.
}
\examples{
\dontrun{

path <- file.path(system.file(package = "eyedroppeR"), "images", "hex.png")

# Run eyedropper and click on 4 colours
pal <- eyedropper(n = 4, path)

pal

}
}
