% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-palette.R
\name{generate_palette}
\alias{generate_palette}
\title{Generate palette}
\usage{
generate_palette(n, prompt)
}
\arguments{
\item{n}{Number of colours in the palette}

\item{prompt}{Prompt}
}
\description{
Generate palette
}
\examples{
# x <- generate_palette(4, 'a trail in a dense rainforest')
}
