% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-palette.R
\name{generate_palette}
\alias{generate_palette}
\title{Generate palette}
\usage{
generate_palette(n, prompt)
}
\arguments{
\item{n}{Number of colours in the palette}

\item{prompt}{Prompt}
}
\value{
A vector of hex codes
}
\description{
⚠️ DEPRECATED: openai has been archived and I have to work out how to generate
images with {ellmer} and the new APIs.
}
\examples{
# x <- generate_palette(4, 'a trail in a dense rainforest')
}
