% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_output}
\alias{make_output}
\title{Makes eyedroppers output}
\usage{
make_output(obj = NULL, pal, img_path)
}
\arguments{
\item{obj}{Output from \code{extract_pal} or \code{eyedropper}}

\item{pal}{Palette}

\item{img_path}{Image path}
}
\value{
ggplot object
}
\description{
Plots the palette and places the image and label over the top.
}
