% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyedropper.R
\name{sort_pal}
\alias{sort_pal}
\title{Manually sort a palette}
\usage{
sort_pal(pal, n = NULL)
}
\arguments{
\item{pal}{Palette. Character vector of hex codes}

\item{n}{Number of colours to choose}
}
\value{
Character vector
}
\description{
The palette is displayed in the plotting window where you can click
the colours in the order you want to sort them. The sorted palette
will be returned.
}
\examples{
\dontrun{
pal <- sample(c('#57364e', '#566f1b', '#97a258', '#cac58b', '#dbedd5'))
sort_pal(pal)
}
}
