% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_pal_auto}
\alias{sort_pal_auto}
\title{Auto palette sort}
\usage{
sort_pal_auto(pal, label, plot_output = FALSE)
}
\arguments{
\item{pal}{Input palette}

\item{label}{Label for the palette.}

\item{plot_output}{Logical. Default \code{FALSE}.}
}
\value{
Returns a character vector of hex codes
}
\description{
Automatically sorts the palette. May not give the desired result. If not you
can run `sort_pal()` to manually sort.
}
\examples{
pal <- sample(colours(), 8)
sort_pal_auto(pal, 'test')
}
