% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swatch}
\alias{swatch}
\title{Swatch of colour palette}
\usage{
swatch(
  pal,
  img = NULL,
  family = "Poppins",
  .padding = 0,
  .radius = 50,
  ncols = NULL
)
}
\arguments{
\item{pal}{Vector of colours}

\item{img}{Image address. Either local file or URL}

\item{family}{Font family}

\item{.padding}{To add whitespace to the top of image}

\item{.radius}{The radius of the feature image. Choose 50 for a circle, less then 50 for rounded square/rectangle.}

\item{ncols}{Number of cols in a row}
}
\value{
html doc
}
\description{
Swatch of colour palette
}
\examples{
url <- "https://github.com/doehm/eyedroppeR/raw/main/dev/images/sunset-south-coast.jpg"
x <- extract_pal(4, url)
swatch(x$pal, url)
}
